requirements/fml_api.txt
# All packages needed for the FML API app.

-r common.txt

djangorestframework==3.15.1
